# digitalib-api-externa/openapi/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API REST EXTERNA de DigitalIB",
    "description" : "Conjunt de Serveis REST de DigitalIB per ser accedits des de l'exterior",
    "contact" : {
      "name" : "Departament de Govern Digital a la Fundació Bit",
      "url" : "http://otae.fundaciobit.org",
      "email" : "otae@fundaciobit.org"
    },
    "license" : {
      "name" : "License Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version" : "1.0.0"
  },
...
  "servers" : [ {
    "url" : "/digitalibapi/externa"
  }, {
    "url" : "http://localhost:8080/digitalibapi/externa"
  }, {
    "url" : "https://dev.caib.es/digitalibapi/externa"
  }, {
    "url" : "https://proves.caib.es/digitalibapi/externa"
  }, {
    "url" : "https://se.caib.es/digitalibapi/externa"
  }, {
    "url" : "https://www.caib.es/digitalibapi/externa"
  } ],
...
  "paths" : {
    "/public/dadesobertes/estadistiques/consulta" : {
      "get" : {
        "tags" : [ "Estadistiques" ],
        "summary" : "Retorna estadístiques de les peticions a DigitalIB",
        "operationId" : "consulta",
        "parameters" : [ {
          "name" : "startdate",
          "in" : "query",
          "description" : "Data d'inici de la consulta. Opcional. Format dd-mm-yyyy",
          "schema" : {
            "pattern" : "^(?:(?:31(-)(?:0?[13578]|1[02]))\\1|(?:(?:29|30)(-)(?:0?[13-9]|1[0-2])\\2))(?:(?:1[6-9]|[2-9]\\d)?\\d{2})$|^(?:29(-)0?2\\3(?:(?:(?:1[6-9]|[2-9]\\d)?(?:0[48]|[2468][048]|[13579][26])|(?:(?:16|[2468][048]|[3579][26])00))))$|^(?:0?[1-9]|1\\d|2[0-8])(-)(?:(?:0?[1-9])|(?:1[0-2]))\\4(?:(?:1[6-9]|[2-9]\\d)?\\d{2})$",
            "type" : "string"
          },
          "example" : "24-07-2023"
        }, {
          "name" : "enddate",
          "in" : "query",
          "description" : "Data final de la consulta. Opcional. Format dd-mm-yyyy",
          "schema" : {
            "pattern" : "^(?:(?:31(-)(?:0?[13578]|1[02]))\\1|(?:(?:29|30)(-)(?:0?[13-9]|1[0-2])\\2))(?:(?:1[6-9]|[2-9]\\d)?\\d{2})$|^(?:29(-)0?2\\3(?:(?:(?:1[6-9]|[2-9]\\d)?(?:0[48]|[2468][048]|[13579][26])|(?:(?:16|[2468][048]|[3579][26])00))))$|^(?:0?[1-9]|1\\d|2[0-8])(-)(?:(?:0?[1-9])|(?:1[0-2]))\\4(?:(?:1[6-9]|[2-9]\\d)?\\d{2})$",
            "type" : "string"
          },
          "example" : "31-12-2024"
        }, {
          "name" : "appname",
          "in" : "query",
          "description" : "Username de l'app. Opcional. Només es pot elegir appname o usrname a la vegada",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "usrname",
          "in" : "query",
          "description" : "Username de la persona. Opcional. Només es pot elegir appname o usrname a la vegada",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "estat",
          "in" : "query",
          "description" : "Estat en que es troba la transacció.  Opcional. Valors possibles són: \"EXPIRAT\",\"CANCELAT\",\"ERROR\",\"ID\",\"ENPROGRES\",\"OK\"",
          "schema" : {
            "pattern" : "^(EXPIRAT|CANCELAT|ERROR|ID|ENPROGRES|OK)$",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Operació realitzada correctament",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Estadistiques"
                }
              }
            }
          },
          "404" : {
            "description" : "Paràmetres incorrectes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "500" : {
            "description" : "Error no controlat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Estadistiques" : {
        "type" : "object",
        "properties" : {
          "resultat" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EstadistiquesTransaccio"
            }
          }
        }
      },
      "EstadistiquesTransaccio" : {
        "type" : "object",
        "properties" : {
          "transaccioID" : {
            "type" : "integer",
            "format" : "int64"
          },
          "transaccioMultipleID" : {
            "type" : "integer",
            "format" : "int64"
          },
          "funcionariUsername" : {
            "type" : "string"
          },
          "appname" : {
            "type" : "string"
          },
          "usrname" : {
            "type" : "string"
          },
          "fitxerMidaBytes" : {
            "type" : "integer",
            "format" : "int64"
          },
          "color" : {
            "type" : "string"
          },
          "resolucio" : {
            "type" : "integer",
            "format" : "int32"
          },
          "midaPaper" : {
            "type" : "string"
          },
          "dataCapturaISO8601" : {
            "type" : "string"
          },
          "estat" : {
            "type" : "string"
          },
          "codiDir3" : {
            "type" : "string"
          },
          "configuracioGrupNom" : {
            "type" : "string"
          },
          "idiomaDocument" : {
            "type" : "string"
          },
          "duplex" : {
            "type" : "boolean"
          },
          "missatgeError" : {
            "type" : "string"
          },
          "origen" : {
            "type" : "string"
          },
          "tipusDocumental" : {
            "type" : "string"
          }
        }
      },
      "RestExceptionInfo" : {
        "required" : [ "code", "errorMessage" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "description" : "Codi de HTTP de l'error. Veure https://en.wikipedia.org/wiki/List_of_HTTP_status_codes.",
            "format" : "int32"
          },
          "errorMessage" : {
            "type" : "string",
            "description" : "Missatge de l'error"
          },
          "stackTrace" : {
            "type" : "string",
            "description" : "Stacktrace de l'excepció si n'hi hagués."
          },
          "causeException" : {
            "type" : "string",
            "description" : "Tipus de l'excepció origen (cause) si n'hi hagués."
          },
          "causeStackTrace" : {
            "type" : "string",
            "description" : "Stacktrace de l'excepció origen (cause) si n'hi hagués."
          }
        },
        "description" : "Estructura de dades utilitzada per passar informació d'un error"
      },
...
    },
    "securitySchemes" : {
      "BasicAuth" : {
        "type" : "http",
        "scheme" : "basic"
      }
    }
  }
}